requirements.txt
# Collecte de données
atproto>=0.0.46
requests>=2.31

# NLP & Machine Learning
transformers>=4.36
torch>=2.1
scikit-learn>=1.3
spacy>=3.7
nltk>=3.8

# Analyse émotionnelle
vaderSentiment>=3.3

# Base de données
psycopg2-binary>=2.9
sqlalchemy>=2.0

# Dashboard
streamlit>=1.29
plotly>=5.18
pandas>=2.1

# Monitoring énergétique (Green IT)
codecarbon>=2.3

# Tests
pytest>=7.4
pytest-cov>=4.1

# Utilitaires
python-dotenv>=1.0
loguru>=0.7
